make Modal work
